+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.2] 2024-06-03
 
 - Add projectname in parentheses to search result on todos/tickets
 - Expand filter results by tolowercase'ing, because the result from the leantime has type set to be either "Task" and "task".